flask-dashboard/requirements.txt
flask==2.3.3
cassandra-driver==3.28.0
flask-cors==4.0.0
python-dateutil==2.8.2